server/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
flask-pymongo==2.3.0
bcrypt==4.0.1
PyJWT==2.8.0
scikit-learn==1.3.2
nltk==3.9.0
requests==2.32.0
python-dotenv==1.0.0
gunicorn==21.2.0
